refactor: tela adicionar meta modularizada
