+++ frontend/lib/features/progress_record/screens/progress_record_screen.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'goal_settings_screen.dart';
-import '../widgets/stat_card.dart';
-import '../widgets/period_chip.dart';
-import '../widgets/chart_type_icon.dart';
-import '../widgets/show_history_bottom_sheet.dart';
+import '../../listHabits/widgets/stat_card.dart';
+import '../../listHabits/widgets/period_chip.dart';
+import '../../listHabits/widgets/chart_type_icon.dart';
+import '../../listHabits/widgets/show_history_bottom_sheet.dart';
 
 class ProgressRecordScreen extends StatefulWidget {
   final String habitName;
@@ -272,8 +272,8 @@ class _ProgressRecordScreenState extends State<ProgressRecordScreen> {
                     ),
                     const SizedBox(height: 16),
                     Wrap(
-                        spacing: 8,
-                        runSpacing: 8,
+                      spacing: 8,
+                      runSpacing: 8,
                       children: [
                         PeriodChip(
                           label: 'Semana',
